requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
deta==1.1.0
fastapi==0.82.0
h11==0.13.0
idna==3.3
MarkupSafe==2.1.1
Pillow==9.2.0
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
qrcode==7.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.3
